More prop cleanup
