+++ packages/react/src/components/CopyIcon/types.ts
@@ -1,0 +1,12 @@
+import type { CSSProperties } from "react";
+
+export type CopyIconProps = {
+  /**
+   * Extra classes to apply to the icon
+   */
+  className?: string;
+  /**
+   * Extra styles to apply to the icon
+   */
+  style?: CSSProperties;
+};
