+++ packages/react/src/components/EmbedProvider/types.ts
@@ -1,0 +1,1 @@
+export type EmbedProviderProps = Record<string, never>;
