+++ packages/react/src/components/ResetIcon/types.ts
@@ -1,0 +1,12 @@
+import type { CSSProperties } from "react";
+
+export type ResetIconProps = {
+  /**
+   * Extra classes to apply to the icon
+   */
+  className?: string;
+  /**
+   * Extra styles to apply to the icon
+   */
+  style?: CSSProperties;
+};
