+++ packages/react/src/components/Tag/types.ts
@@ -1,5 +1,20 @@
 import type { PropsWithChildren } from "react";
 
-export type TagProps = PropsWithChildren<{ slot: "tag" }>;
-export type TagTitleProps = PropsWithChildren<{ slot: "title" }>;
-export type TagDescriptionProps = PropsWithChildren<{ slot: "description" }>;
+export type TagProps = PropsWithChildren<{
+  /**
+   * The slot for the tag, always "tag"
+   */
+  slot: "tag";
+}>;
+export type TagTitleProps = PropsWithChildren<{
+  /**
+   * The slot for the tag title, always "title"
+   */
+  slot: "title";
+}>;
+export type TagDescriptionProps = PropsWithChildren<{
+  /**
+   * The slot for the tag description, always "description"
+   */
+  slot: "description";
+}>;
